requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.3.3
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.136.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.1
fastar==0.11.0
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.27
python-telegram-bot==22.1
pytz==2025.2
PyYAML==6.0.3
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.38
starlette==0.46.2
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
